requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.0
crispy-bootstrap4==2024.10
Django==4.2.17
django-bootstrap-v5==1.0.11
django-ckeditor==6.7.2
django-crispy-forms==2.3
django-jazzmin==3.0.1
django-js-asset==2.2.0
djangorestframework==3.15.2
idna==3.10
pillow==11.0.0
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
